+++ binary-ninja/shuriken-service/inc/shurikenservice_internal.h
@@ -17,7 +17,7 @@ public:
     BinaryNinja::BinaryView *getBinaryView() { return m_bv; };
 
     const shurikenapi::IDex &getParser() const override;
-    uint32_t getStringOffset(uint32_t id) const override;
+    uint32_t getStringOffset(int64_t id) const override;
 
 private:
     BinaryNinja::Ref<BinaryNinja::Function> buildMethod(const shurikenapi::IClassMethod &method);
@@ -25,7 +25,7 @@ private:
 
     BinaryNinja::BinaryView *m_bv;
     std::unique_ptr<shurikenapi::IDex> m_parser;
-    std::unordered_map<uint32_t, uint32_t> m_stringMap;
+    std::unordered_map<int64_t, uint32_t> m_stringMap;
 };
 
 class ShurikenService : public IShurikenService {
